@@ -72,7 +72,6 @@
                                     <value>${rvn.config}</value>
                                 </systemProperty>
                             </systemProperties>
-
                         </configuration>
                     </plugin>
                 </plugins>
@@ -116,7 +115,7 @@
                                             <version>2.0.6</version>
                                         </requireMavenVersion>
                                         <requireJavaVersion>
-                                            <version>1.8</version>
+                                            <version>9</version>
                                         </requireJavaVersion>
                                         <!-- <requireOS> <family>unix</family> </requireOS> -->
 
@@ -125,6 +124,7 @@
                             </execution>
                         </executions>
                     </plugin>
+                    <!--
                     <plugin>
                         <groupId>org.codehaus.mojo</groupId>
                         <artifactId>animal-sniffer-maven-plugin</artifactId>
@@ -153,6 +153,7 @@
                             </execution>
                         </executions>
                     </plugin>
+            -->
                     <plugin>
                         <groupId>com.oracle.substratevm</groupId>
                         <artifactId>native-image-maven-plugin</artifactId>
@@ -378,6 +379,18 @@
             <scope>test</scope>
         </dependency>
                 -->
+        <dependency>
+            <groupId>org.graalvm.truffle</groupId>
+            <artifactId>truffle-api</artifactId>
+            <version>21.1.0</version>
+            <type>jar</type>
+        </dependency>
+        <dependency>
+            <groupId>javax</groupId>
+            <artifactId>javaee-api</artifactId>
+            <version>7.0</version>
+            <type>jar</type>
+        </dependency>
     </dependencies>
     <distributionManagement
         xmlns:mvn="http://maven.apache.org/POM/4.0.0">
@@ -406,8 +419,8 @@
     </scm>
 
     <properties>
-        <maven.compiler.source>8</maven.compiler.source>
-        <maven.compiler.target>8</maven.compiler.target>
+        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.target>11</maven.compiler.target>
     </properties>
 
 </project>
